Menu Inventory System(Queues, Sets, Maps)
This program is a restaurant inventory System. The user can add food items along with their ingredients and the quantity of each ingredient. The user can also see the ingredients needed for a food item, update the ingredient quantities based on the ingredients used, display this information on the user's screen